#55 Refactor: Schemas 각 도메인 하위로 분배
