--- backend/app/api/schemas/default.py
@@ -1,8 +1,0 @@
-from pydantic import BaseModel
-
-
-class ResourceId(BaseModel):
-    id: int
-
-    class Config:
-        from_attributes = True
