+++ backend/app/api/v1/search/search.py
@@ -1,8 +1,7 @@
 from fastapi import Depends, status, APIRouter, Request, HTTPException
 from sqlalchemy.orm.session import Session
 
-from app.api.schemas import search_schema
-from app.api.v1.search import search_utils
+from app.api.v1.search import search_utils, schema
 from app import model
 from app.database import get_db
 
@@ -11,7 +10,7 @@ router = APIRouter(tags=["Search"])
 
 # Pydantic 모델을 query parameter로 받을 때는 Depends() 사용할 것
 @router.get("/hashtag", status_code=status.HTTP_200_OK)
-async def get_hashtag_result(data: search_schema.SearchHashTag = Depends(), db: Session = Depends(get_db)):
+async def get_hashtag_result(data: schema.SearchHashTag = Depends(), db: Session = Depends(get_db)):
     """
     검색한 해시태그에 해당하는 결과 return
     """
@@ -28,7 +27,7 @@ async def get_hashtag_result(data: search_schema.SearchHashTag = Depends(), db: 
 
 
 @router.get("/nickname", status_code=status.HTTP_200_OK)
-async def get_name_result(data: search_schema.SearchName = Depends(), db: Session = Depends(get_db)):
+async def get_name_result(data: schema.SearchName = Depends(), db: Session = Depends(get_db)):
     """
     검색한 사용자 닉네임에 해당하는 결과 return
     """
@@ -44,9 +43,7 @@ async def get_name_result(data: search_schema.SearchName = Depends(), db: Sessio
 
 
 @router.get("/main", status_code=status.HTTP_200_OK)
-async def get_main_result(
-    request: Request, data: search_schema.SearchFollower = Depends(), db: Session = Depends(get_db)
-):
+async def get_main_result(request: Request, data: schema.SearchFollower = Depends(), db: Session = Depends(get_db)):
     decoded_dict = request.state.decoded_dict
     if decoded_dict:
         return await search_utils.return_follow_posts(
